pyproject.toml
[project]
name = "strava-analysis"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "jupyter>=1.1.1",
    "pandas>=2.3.3",
    "polyline>=2.0.3",
    "psycopg2-binary>=2.9.11",
    "python-dotenv>=1.1.1",
    "requests-oauthlib>=2.0.0",
    "rich>=14.2.0",
    "stravalib>=2.4",
    "streamlit>=1.50.0",
    "watchdog>=6.0.0",
]
